Refactor: Align frontend architecture with best practices
This commit addresses several inconsistencies and opportunities for improvement in the frontend codebase to align with the project's defined architecture and best practices.

Key changes include:

1.  **Centralized Zustand Store (`agentStore`):**
    *   Moved `agentStore.ts` from `client/src/components/features/agents-list/store/` to the designated `client/src/store/` directory.
    *   Updated all import paths.

2.  **Consolidated Agent Types:**
    *   Moved type definitions from `client/src/components/agents/types.ts` to `client/src/types/agent.ts`.
    *   Removed the redundant types file.

3.  **Decoupled `toolService.ts`:**
    *   Removed direct Zustand store (`useToolStore`) dependency from `toolService.ts`.
    *   `fetchTools` now returns mock data directly.
    *   State-mutating service methods (`addTool`, `updateTool`, `deleteTool`) that were manipulating the store directly have been removed from the service layer.
    *   Associated tests in `toolService.test.ts` were updated.

4.  **Eliminated Duplicate `apiClient.ts`:**
    *   Removed the redundant `client/src/apiClient.ts`.
    *   Ensured all usages correctly import from `client/src/api/apiClient.ts`.

5.  **Unified UI Components:**
    *   Consolidated `dialog.tsx` and `dialog-new.tsx` into `dialog.tsx`, keeping the more complete version and merging necessary type exports.
    *   Consolidated `theme-provider.tsx` (root) and `theme-provider.tsx` (UI) into the UI version, which offers full theme switching functionality.

6.  **Refactored `AgentsDashboard.tsx`:**
    *   Moved `mockAgents` data to `client/src/data/mocks/mock-dashboard-agents.ts`.
    *   Removed local definitions of `dashboardAgentTypeLabels`. The component now uses centralized utilities.

7.  **Centralized Agent Utilities:**
    *   Moved `agentTypeLabels` and `getAgentTypeColor` from `AgentListItem.tsx` to `client/src/lib/agent-utils.ts`.
    *   Updated `AgentListItem.tsx` and `AgentsDashboard.tsx` to use these centralized utilities.

8.  **Standardized Page Component Naming:**
    *   Renamed page components in `client/src/pages/` and its subdirectories to the `NomeDaPaginaPage.tsx` convention (e.g., `Dashboard.tsx` to `DashboardPage.tsx`).
    *   Updated import paths in `client/src/routes.tsx`.

These changes improve the maintainability, consistency, and architectural integrity of the frontend application.
