+++ client/src/api/index.ts
@@ -4,4 +4,4 @@ export * from './dashboardService'
 export * from './memoryService'
 export * from './toolService'
 export * from './secretService'
-export { default as apiClient } from './apiClient'
+export { default as apiClient } from '@/api/apiClient'
