+++ client/src/api/secretService.ts
@@ -1,4 +1,4 @@
-import apiClient from '../apiClient'
+import apiClient from '@/api/apiClient'
 import { SecretDTO } from '@/types/api'
 
 export const storeSecret = async (payload: SecretDTO): Promise<void> => {
