--- client/src/components/ui/dialog-new.tsx
@@ -1,178 +1,0 @@
-import * as React from 'react';
-import * as DialogPrimitive from '@radix-ui/react-dialog';
-import { X } from 'lucide-react';
-import { cn } from '@/lib/utils';
-
-// Base Dialog Components
-export const Dialog = DialogPrimitive.Root;
-export const DialogTrigger = DialogPrimitive.Trigger;
-
-// Dialog Portal
-interface DialogPortalProps {
-  children: React.ReactNode;
-  className?: string;
-}
-
-const DialogPortal: React.FC<DialogPortalProps> = ({
-  children,
-  className,
-  ...props
-}) => (
-  <DialogPrimitive.Portal>
-    <div 
-      className={cn(
-        'fixed inset-0 z-50 flex items-start justify-center sm:items-center', 
-        className
-      )} 
-      {...props}
-    >
-      {children}
-    </div>
-  </DialogPrimitive.Portal>
-);
-
-DialogPortal.displayName = 'DialogPortal';
-
-// Dialog Overlay
-interface DialogOverlayProps 
-  extends React.ComponentPropsWithoutRef<typeof DialogPrimitive.Overlay> {}
-
-const DialogOverlay = React.forwardRef<
-  React.ElementRef<typeof DialogPrimitive.Overlay>,
-  DialogOverlayProps
->(({ className, ...props }, ref) => (
-  <DialogPrimitive.Overlay
-    ref={ref}
-    className={cn(
-      'fixed inset-0 z-50 bg-background/80 backdrop-blur-sm',
-      'data-[state=open]:animate-in data-[state=closed]:animate-out',
-      'data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0',
-      className
-    )}
-    {...props}
-  />
-));
-
-DialogOverlay.displayName = DialogPrimitive.Overlay.displayName;
-
-// Dialog Content
-interface DialogContentProps 
-  extends React.ComponentPropsWithoutRef<typeof DialogPrimitive.Content> {}
-
-const DialogContent = React.forwardRef<
-  React.ElementRef<typeof DialogPrimitive.Content>,
-  DialogContentProps
->(({ className, children, ...props }, ref) => (
-  <DialogPortal>
-    <DialogOverlay />
-    <DialogPrimitive.Content
-      ref={ref}
-      className={cn(
-        'fixed z-50 grid w-full gap-4 rounded-b-lg border bg-background p-6 shadow-lg',
-        'sm:max-w-lg sm:rounded-lg',
-        'data-[state=open]:animate-in data-[state=closed]:animate-out',
-        'data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0',
-        'data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95',
-        'data-[state=closed]:slide-out-to-left-1/2 data-[state=closed]:slide-out-to-top-[48%]',
-        'data-[state=open]:slide-in-from-left-1/2 data-[state=open]:slide-in-from-top-[48%]',
-        className
-      )}
-      {...props}
-    >
-      {children}
-      <DialogPrimitive.Close className="absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none">
-        <X className="h-4 w-4" />
-        <span className="sr-only">Close</span>
-      </DialogPrimitive.Close>
-    </DialogPrimitive.Content>
-  </DialogPortal>
-));
-
-DialogContent.displayName = DialogPrimitive.Content.displayName;
-
-// Dialog Header
-const DialogHeader: React.FC<React.HTMLAttributes<HTMLDivElement>> = ({
-  className,
-  ...props
-}) => (
-  <div
-    className={cn(
-      'flex flex-col space-y-1.5 text-center sm:text-left',
-      className
-    )}
-    {...props}
-  />
-);
-
-DialogHeader.displayName = 'DialogHeader';
-
-// Dialog Footer
-const DialogFooter: React.FC<React.HTMLAttributes<HTMLDivElement>> = ({
-  className,
-  ...props
-}) => (
-  <div
-    className={cn(
-      'flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2',
-      className
-    )}
-    {...props}
-  />
-);
-
-DialogFooter.displayName = 'DialogFooter';
-
-// Dialog Title
-interface DialogTitleProps 
-  extends React.ComponentPropsWithoutRef<typeof DialogPrimitive.Title> {}
-
-const DialogTitle = React.forwardRef<
-  React.ElementRef<typeof DialogPrimitive.Title>,
-  DialogTitleProps
->(({ className, ...props }, ref) => (
-  <DialogPrimitive.Title
-    ref={ref}
-    className={cn(
-      'text-lg font-semibold leading-none tracking-tight',
-      className
-    )}
-    {...props}
-  />
-));
-
-DialogTitle.displayName = DialogPrimitive.Title.displayName;
-
-// Dialog Description
-interface DialogDescriptionProps 
-  extends React.ComponentPropsWithoutRef<typeof DialogPrimitive.Description> {}
-
-const DialogDescription = React.forwardRef<
-  React.ElementRef<typeof DialogPrimitive.Description>,
-  DialogDescriptionProps
->(({ className, ...props }, ref) => (
-  <DialogPrimitive.Description
-    ref={ref}
-    className={cn('text-sm text-muted-foreground', className)}
-    {...props}
-  />
-));
-
-DialogDescription.displayName = DialogPrimitive.Description.displayName;
-
-export {
-  DialogPortal,
-  DialogOverlay,
-  DialogContent,
-  DialogHeader,
-  DialogFooter,
-  DialogTitle,
-  DialogDescription,
-};
-
-export type {
-  DialogPortalProps,
-  DialogOverlayProps,
-  DialogContentProps,
-  DialogTitleProps,
-  DialogDescriptionProps,
-};
