+++ client/src/data/mocks/mock-dashboard-agents.ts
@@ -1,0 +1,54 @@
+import { Agent } from '@/types/agents'; // Ensure Agent type is imported
+
+// Mock data - replace with actual API calls
+export const mockAgents: Agent[] = [
+  {
+    id: '1',
+    name: 'Customer Support Bot',
+    description: 'Handles customer inquiries and support tickets',
+    type: 'llm',
+    model: 'gemini-1.5-pro',
+    temperature: 0.7,
+    maxTokens: 1024,
+    topP: 1,
+    topK: 40,
+    stopSequences: [],
+    frequencyPenalty: 0,
+    presencePenalty: 0,
+    instruction: 'You are a helpful customer support assistant. Be polite and professional.',
+    systemPrompt: '## Guidelines\n- Always verify customer information\n- Escalate complex issues to a human agent\n- Provide clear next steps',
+    version: '1.0.0',
+    isPublic: false,
+    tags: ['support', 'customer-service'],
+    createdAt: '2024-01-01T00:00:00Z',
+    updatedAt: '2024-01-01T00:00:00Z',
+  },
+  {
+    id: '2',
+    name: 'Document Analysis Workflow',
+    description: 'Processes and analyzes documents in a sequential workflow',
+    type: 'sequential',
+    agents: ['1', '3'],
+    maxSteps: 5,
+    stopOnError: true,
+    version: '1.0.0',
+    isPublic: true,
+    tags: ['document', 'workflow', 'automation'],
+    createdAt: '2024-01-15T00:00:00Z',
+    updatedAt: '2024-01-15T00:00:00Z',
+  },
+  {
+    id: '3',
+    name: 'Data Processing Pipeline',
+    description: 'Processes data in parallel for better performance',
+    type: 'parallel',
+    agents: ['1', '4'],
+    maxConcurrent: 3,
+    timeoutMs: 30000,
+    version: '1.0.0',
+    isPublic: true,
+    tags: ['data', 'pipeline', 'performance'],
+    createdAt: '2024-02-01T00:00:00Z',
+    updatedAt: '2024-02-01T00:00:00Z',
+  },
+];
