+++ client/src/pages/MemoryPage.tsx
