+++ client/src/pages/agents/AgentsTemplatesPage.tsx
